수정

+++ MC_backend/src/main/java/com/memory/connect/model/place/repository/PlaceRepository.java
@@ -3,5 +3,5 @@ package com.memory.connect.model.place.repository;
 import com.memory.connect.model.place.entity.Place;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PlaceRepository extends JpaRepository<Place, Long> {
+public interface PlaceRepository extends JpaRepository<Place, Integer> {
 }
